Move PathConnectDetector to scene-controller; move connectContours to path-functions
